refactor(factorio-api): separate auth logic into submodule
