+++ crates/cli/src/error.rs
@@ -7,8 +7,11 @@ pub enum Error {
     #[error(transparent)]
     RonSpanned(#[from] ron::de::SpannedError),
 
-    #[error("Error from factorio_api: {0:?}")]
-    FactorioApi(factorio_api::ErrorResponse),
+    #[error("Error logging in: {0:?}")]
+    FactorioLogin(factorio_api::auth::LoginError),
+    #[error("Error retrieving mod release info: {0:?}")]
+    FactorioApi(String),
+
     #[error("No matching release found for mod {0} with version {1}")]
     NoMatchingRelease(String, String),
 }
