+++ crates/factorio-api/src/lib.rs
@@ -1,20 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-/// The response returned by the Factorio Mod Portal API when logging in using the Web
-/// Authentication API.
-/// The format of the request is taken from https://wiki.factorio.com/Web_authentication_API
-#[derive(Debug, Deserialize, Serialize)]
-pub struct LoginResponse {
-    pub token: String,
-    pub username: String,
-}
-
-/// An error returned when calling the Factorio Mod Portal API.
-#[derive(Debug, Deserialize, Serialize)]
-pub struct ErrorResponse {
-    error: String,
-    message: String,
-}
+pub mod auth;
 
 /// The response returned by the Factorio Mod Portal API when requesting a mod's info.
 #[derive(Debug, Deserialize, Serialize)]
